+++ python/clinical_asl_pipeline/asl_t1_from_m0.py
@@ -39,19 +39,26 @@ def asl_t1_from_m0(subject, context_tag):
     # Use a shorter alias for context_data
     context_data = subject[context_tag]     
     
-    logging.info('ASL T1-from-M0 computation started')
+    LookLocker_correction_factor_perPLD = context_data['LookLocker_correction_factor_perPLD']
+    T1fromM0_path = context_data['T1fromM0_path']
+    templateNIFTI_path =context_data['templateNIFTI_path']
+    M0_allPLD = context_data['M0_allPLD']
+    brainmask = context_data['mask'] 
+    PLDS =  context_data['PLDS']
+    TR = context_data['TR']
 
     # Remove the Look-Locker correction (by multiplication) to compute the T1w profile (vectorized)
-    M0_allPLD_noLLcorr = context_data['M0_allPLD'] * context_data['LookLocker_correction_factor_perPLD'][None, None, None, :]
+    M0_allPLD_noLLcorr = M0_allPLD * LookLocker_correction_factor_perPLD[None, None, None, :]
+    logging.info('ASL T1-from-M0 computation started')
 
     # Compute T1w image from multi-PLD M0 and save nifti
     logging.info('Create T1w image from multiPLD M0')   
 
-    T1fromM0 = asl_t1_from_m0_compute(M0_allPLD_noLLcorr, context_data['mask'], context_data['PLDS'])
-    save_data_nifti(T1fromM0, context_data['T1fromM0_path'], context_data['templateNIFTI_path'], 1, [0, 500], context_data['TR'])
+    T1fromM0 = asl_t1_from_m0_compute(M0_allPLD_noLLcorr, brainmask, PLDS)
+    save_data_nifti(T1fromM0, T1fromM0_path, templateNIFTI_path, 1, [0, 500], TR)
 
     # Final T1fromM0 load
-    context_data['T1fromM0'] = nib.load(context_data['T1fromM0_path']).get_fdata()
+    context_data['T1fromM0'] = nib.load(T1fromM0_path).get_fdata()
     return subject
 
 def asl_t1_from_m0_compute(DATA4D, MASK, TIMEARRAY):